# docs/swagger.yml
basePath: /api/v1
consumes:
    - application/json
definitions:
    BaseResponseBody:
        description: Contains error and message fields.
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
        required:
            - error
            - message
        title: BaseResponseBody represents the base response body for API.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    Comment:
        description: Comment represents the comment for this application
        properties:
            content:
                description: Comment content
                type: string
                x-go-name: Content
            created_at:
                description: The time the comment was created
                format: date-time
                type: string
                x-go-name: CreatedAt
            id:
                description: The id for this comment
                format: uuid
                type: string
                x-go-name: ID
            liked_by_me:
                type: boolean
                x-go-name: LikedByMe
            likes:
                description: Number of likes
                format: int64
                type: integer
                x-go-name: Likes
            updated_at:
                description: The time the comment was updated
                format: date-time
                type: string
                x-go-name: UpdatedAt
            user:
                $ref: '#/definitions/User'
        required:
            - id
            - content
            - created_at
            - updated_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
    CommentAddRequestBody:
        properties:
            content:
                type: string
                x-go-name: Content
        required:
            - content
        title: CommentAddRequestBody includes the content of the comment.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
    CommentUpdateRequestBody:
        properties:
            content:
                type: string
                x-go-name: Content
        required:
            - content
        title: CommentUpdateRequestBody includes the content of the comment.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
    GetCommentResponseBody:
        description: |-
            GetCommentResponseBody includes the signle comment for a post
            based on given post and comment ID.
        properties:
            comment:
                $ref: '#/definitions/Comment'
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
        required:
            - error
            - message
            - comment
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
...
    GetPostResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            post:
                $ref: '#/definitions/Post'
        required:
            - error
            - message
            - post
        title: GetPostResponseBody includes the signle post for a given ID.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    GetPostsResponseBody:
        properties:
            count:
                format: int64
                type: integer
                x-go-name: Count
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            posts:
                items:
                    $ref: '#/definitions/Post'
                type: array
                x-go-name: Posts
        required:
            - error
            - message
            - count
            - posts
        title: GetPostsResponseBody includes the slice of posts.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
...
    GetUserResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            user:
                $ref: '#/definitions/User'
        required:
            - error
            - message
            - user
        title: GetUserResponseBody includes the signle user for a given ID.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
...
    Post:
        description: Post represents the post for this application
        properties:
            comments:
                items:
                    $ref: '#/definitions/Comment'
                type: array
                x-go-name: Comments
            content:
                description: The url to the post content
                type: string
                x-go-name: Content
            created_at:
                description: The time the post was created
                format: date-time
                type: string
                x-go-name: CreatedAt
            description:
                description: Post description
                type: string
                x-go-name: Description
            id:
                description: The id for this post
                format: uuid
                type: string
                x-go-name: ID
            liked_by_me:
                type: boolean
                x-go-name: LikedByMe
            likes:
                description: Number of likes
                format: int64
                type: integer
                x-go-name: Likes
            user:
                $ref: '#/definitions/User'
        required:
            - id
            - content
            - description
            - created_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
...
    UploadDataResponseBody:
        properties:
            error:
                example: false
                type: boolean
                x-go-name: Error
            message:
                x-go-name: Message
            path:
                type: string
                x-go-name: Path
        required:
            - error
            - message
            - path
        title: UploadDataResponseBody includes the id of uploaded data.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/responses
    User:
        description: User represents the user for this application
        properties:
            avatar_url:
                description: URL to user avatar
                type: string
                x-go-name: AvatarURL
            created_at:
                description: The time the user was registered
                format: date-time
                type: string
                x-go-name: CreatedAt
            email:
                description: The email for this user
                type: string
                x-go-name: Email
            id:
                description: The id for this user
                format: uuid
                type: string
                x-go-name: ID
            name:
                description: The name for this user
                type: string
                x-go-name: Name
            updated_at:
                description: The time the user was updated
                format: date-time
                type: string
                x-go-name: UpdatedAt
            username:
                description: The username for this user
                type: string
                x-go-name: Username
        required:
            - id
            - email
            - username
            - created_at
            - updated_at
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/models
    UserUpdateRequestBody:
        properties:
            email:
                maxLength: 255
                minLength: 6
                type: string
                x-go-name: Email
            name:
                maxLength: 32
                type: string
                x-go-name: Name
            password:
                maxLength: 256
                minLength: 8
                type: string
                x-go-name: Password
            username:
                maxLength: 32
                minLength: 1
                type: string
                x-go-name: Username
        title: UserUpdateRequestBody includes the new email, username, name or password for the user.
        type: object
        x-go-package: github.com/MangriMen/Diverse-Back/internal/parameters
host: localhost:3040
info:
    title: Diverse API.
    version: 0.0.1
paths:
    /data:
        post:
            description: Returns the ID of uploaded data
            operationId: uploadData
            parameters:
                - in: formData
                  name: file
                  type: file
                  x-go-name: File
            responses:
                "200":
                    $ref: '#/responses/UploadDataResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - Data
...
    /posts/{post}:
        delete:
            description: Delete post by ID
            operationId: deletePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            produces:
                - application/json
            responses:
                "204":
                    $ref: '#/responses/DeletePostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            schemes:
                - http
                - https
            security:
                - bearerAuth: []
            tags:
                - Post
        get:
            description: Returns the post by given ID
            operationId: getPost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "200":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        patch:
            description: Update post by ID with given fields
            operationId: updatePost
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
            responses:
                "201":
                    $ref: '#/responses/GetPostResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
    /posts/{post}/comments:
        get:
            description: Returns a list of post comments
            operationId: getComments
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - format: uuid
                  in: query
                  name: last_seen_comment_id
                  type: string
                  x-go-name: LastSeenCommentID
                - description: nolint:lll
                  format: date-time
                  in: query
                  name: last_seen_comment_created_at
                  type: string
                  x-go-name: LastSeenCommentCreatedAt
                - format: int64
                  in: query
                  maximum: 50
                  minimum: 1
                  name: count
                  required: true
                  type: integer
                  x-go-name: Count
            responses:
                "200":
                    $ref: '#/responses/GetPostsResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        post:
            description: Add comment to the given post
            operationId: addComment
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CommentAddRequestBody'
            responses:
                "201":
                    $ref: '#/responses/CreateUpdateCommentResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
    /posts/{post}/comments/{comment}:
        delete:
            description: Delete comment by comment ID with given post ID
            operationId: deleteComment
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - format: uuid
                  in: path
                  name: comment
                  required: true
                  type: string
                  x-go-name: Comment
            produces:
                - application/json
            responses:
                "204":
                    $ref: '#/responses/DeleteCommentResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            schemes:
                - http
                - https
            security:
                - bearerAuth: []
            tags:
                - Post
        patch:
            description: Update comment content by comment ID with given post ID
            operationId: updateComment
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - format: uuid
                  in: path
                  name: comment
                  required: true
                  type: string
                  x-go-name: Comment
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CommentUpdateRequestBody'
            responses:
                "201":
                    $ref: '#/responses/CreateUpdateCommentResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
    /posts/{post}/comments/{comment}/like:
        delete:
            description: Unset like to the comment by ID
            operationId: unlikeComment
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - format: uuid
                  in: path
                  name: comment
                  required: true
                  type: string
                  x-go-name: Comment
            responses:
                "201":
                    $ref: '#/responses/GetCommentResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
        post:
            description: Set like to the comment by ID
            operationId: likeComment
            parameters:
                - format: uuid
                  in: path
                  name: post
                  required: true
                  type: string
                  x-go-name: Post
                - format: uuid
                  in: path
                  name: comment
                  required: true
                  type: string
                  x-go-name: Comment
            responses:
                "201":
                    $ref: '#/responses/GetCommentResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - Post
...
    /users/{user}:
        delete:
            description: Delete user by id
            operationId: deleteUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
            produces:
                - application/json
            responses:
                "204":
                    $ref: '#/responses/DeleteUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            schemes:
                - http
                - https
            security:
                - bearerAuth: []
            tags:
                - User
        get:
            description: Returns the user by given id
            operationId: getUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
            responses:
                "200":
                    $ref: '#/responses/GetUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            tags:
                - User
        patch:
            description: Update user by id with given fields
            operationId: updateUser
            parameters:
                - format: uuid
                  in: path
                  name: user
                  required: true
                  type: string
                  x-go-name: User
                - in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/UserUpdateRequestBody'
            responses:
                "201":
                    $ref: '#/responses/UpdateUserResponse'
                default:
                    $ref: '#/responses/ErrorResponse'
            security:
                - bearerAuth: []
            tags:
                - User
...
produces:
    - application/json
responses:
...
    CreateUpdateCommentResponse:
        description: CreateUpdateCommentResponse represents response for successfully create or update comment request.
        schema:
            type: string
...
    DeleteCommentResponse:
        description: DeleteCommentResponse represents response for successfully delete comment request.
    DeletePostResponse:
        description: DeletePostResponse represents response for successfully delete post.
...
    DeleteUserResponse:
        description: DeleteUserResponse represents response for successfully delete user request.
    ErrorResponse:
        description: |-
            ErrorResponse represents the error response body for API.
            Contains error and message fields.
        schema:
            $ref: '#/definitions/BaseResponseBody'
    GetCommentResponse:
        description: GetCommentResponse represent the response retrived on get comment request.
        schema:
            $ref: '#/definitions/GetCommentResponseBody'
...
    GetPostResponse:
        description: GetPostResponse represent the response retrived on get post request.
        schema:
            $ref: '#/definitions/GetPostResponseBody'
    GetPostsResponse:
        description: GetPostsResponse represent the response retrived on get posts request.
        schema:
            $ref: '#/definitions/GetPostsResponseBody'
...
    GetUserResponse:
        description: GetUserResponse represent the response retrived on get user request.
        schema:
            $ref: '#/definitions/GetUserResponseBody'
...
    UpdateUserResponse:
        description: UpdateUserResponse represents response for successfully update user request.
        schema:
            type: string
    UploadDataResponse:
        description: UploadDataResponse represent the response retrived on upload data request.
        schema:
            $ref: '#/definitions/UploadDataResponseBody'
schemes:
    - http
    - https
securityDefinitions:
    bearerAuth:
        description: JWT Authorization header
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"